Arreglado el tipado el rating de VoteRequest

+++ movies/src/main/java/sithmcfly/movies/http/request/VoteRequest.java
@@ -6,17 +6,17 @@ import jakarta.validation.constraints.Min;
 public class VoteRequest {
     @Min(value = 1, message = "La nota no puede ser inferior a 1")
     @Max(value = 10, message = "La nota no puede ser superior a 10")
-    private int rating;
+    private double rating;
     
-    public int getRating() {
+    public double getRating() {
         return rating;
     }
 
-    public void setRating(int rating) {
+    public void setRating(double rating) {
         this.rating = rating;
     }
 
-    public VoteRequest(int rating) {
+    public VoteRequest(double rating) {
         this.rating = rating;
     }
